feat: Add NeuronGraphDataset class
The `NeuronGraphDataset` class is a subclass of `DGLDataset` that provides an interface for loading and
processing a collection of DGL graphs from SWC files. It accepts a `graphs_path` argument specifying the path to
the directory containing the SWC files, and optionally a
`self_loop` argument to add self-loops to the graphs.