fix: include all negative samples in NT-Xent implementation.
The previous implementation of the NT-Xent loss function was not including
all 2N-1 negative samples in the loss calculation. This was due to a mistake
in the creation of the negative mask. This commit fixes the implementation
by using the correct negative mask and changing the argument names to make
them more intuitive.

Changes:
- Rename batch_embeddings to embed_a, and
batch_augmented_embeddings to embed_b
- Use the correct negative mask in loss calculation